server/requirements.txt
arabic-reshaper==2.1.3
asgiref==3.3.4
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.7
defusedxml==0.7.1
Django==3.2.4
django-allauth==0.44.0
django-crispy-forms==1.12.0
django-js-asset==1.2.2
django-mptt==0.12.0
django-widget-tweaks==1.4.8
Faker==8.8.0
future==0.18.2
html5lib==1.1
idna==2.10
install==1.3.4
oauthlib==3.1.1
Pillow==8.2.0
pycparser==2.20
PyJWT==2.1.0
PyPDF2==1.26.0
python-bidi==0.4.2
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2021.1
qrcode==6.1
reportlab==3.5.67
requests==2.25.1
requests-oauthlib==1.3.0
six==1.16.0
sqlparse==0.4.1
text-unidecode==1.3
urllib3==1.26.5
webencodings==0.5.1
xhtml2pdf==0.2.5
